+++ src/Telegram.Bot/Types/InlineQueryResults/InputMessageContent.cs
@@ -1,7 +1,7 @@
 ﻿using Newtonsoft.Json;
 using Newtonsoft.Json.Serialization;
 
-namespace Telegram.Bot.Types.InputMessageContents
+namespace Telegram.Bot.Types.InlineQueryResults
 {
     /// <summary>
     /// This object represents the content of a message to be sent as a result of an <see cref="InlineQuery"/>. Telegram clients currently support the following 4 types <see cref="InputTextMessageContent"/>, <see cref="InputLocationMessageContent"/>, <see cref="InputVenueMessageContent"/>, <see cref="InputContactMessageContent"/>