+++ src/Telegram.Bot/Types/InlineQueryResults/InputTextMessageContent.cs
@@ -2,7 +2,7 @@ using Newtonsoft.Json;
 using Newtonsoft.Json.Serialization;
 using Telegram.Bot.Types.Enums;
 
-namespace Telegram.Bot.Types.InputMessageContents
+namespace Telegram.Bot.Types.InlineQueryResults
 {
     /// <summary>
     /// Represents the content of a text message to be sent as the result of an <see cref="InlineQuery"/>.